0x0A. Configuration management

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -3,4 +3,6 @@
 exec { 'pkill':
   command  => 'pkill -f killmenow',
   path     => '/usr/bin',
+  onlyif   => 'pgrep killmenow',
+  provider => shell,
 }
